+++ modules/compute/src/main/java/org/apache/ignite/internal/compute/executor/ComputeExecutorImpl.java
@@ -22,6 +22,7 @@ import static org.apache.ignite.internal.thread.ThreadOperation.STORAGE_WRITE;
 
 import java.util.concurrent.atomic.AtomicBoolean;
 import org.apache.ignite.Ignite;
+import org.apache.ignite.compute.ComputeException;
 import org.apache.ignite.compute.ComputeJob;
 import org.apache.ignite.compute.JobExecutionContext;
 import org.apache.ignite.compute.task.MapReduceTask;
@@ -40,6 +41,7 @@ import org.apache.ignite.internal.compute.task.TaskExecutionInternal;
 import org.apache.ignite.internal.logger.IgniteLogger;
 import org.apache.ignite.internal.logger.Loggers;
 import org.apache.ignite.internal.thread.IgniteThreadFactory;
+import org.apache.ignite.lang.ErrorGroups.Compute;
 import org.apache.ignite.marshaling.Marshaler;
 import org.jetbrains.annotations.Nullable;
 
@@ -122,7 +124,7 @@ public class ComputeExecutorImpl implements ComputeExecutor {
             return marshaler.unmarshal((byte[]) input);
         }
 
-        return (T) input;
+        throw new ComputeException(Compute.TYPE_CHECK_MARSHALING_ERR, "test");
     }
 
     @Override
