+++ modules/core/src/main/java/org/apache/ignite/internal/type/NativeTypeSpec.java
@@ -38,7 +38,6 @@ import org.jetbrains.annotations.Nullable;
  * <p>An instance of native type provides necessary indirection to read any field as an instance of {@code java.lang.Object} to avoid
  * switching inside the row methods.
  */
-
 public enum NativeTypeSpec {
     /**
      * Native type representing a single-byte signed value.
