+++ modules/runner/src/integrationTest/java/org/apache/ignite/internal/runner/app/client/ItThinClientComputeMarshallingTest.java
@@ -17,28 +17,27 @@
 
 package org.apache.ignite.internal.runner.app.client;
 
-import static java.util.concurrent.CompletableFuture.completedFuture;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import java.io.IOException;
 import java.util.Comparator;
 import java.util.List;
-import java.util.concurrent.CompletableFuture;
 import java.util.stream.Collectors;
 import org.apache.ignite.Ignite;
 import org.apache.ignite.client.IgniteClient;
-import org.apache.ignite.compute.ComputeJob;
 import org.apache.ignite.compute.IgniteCompute;
 import org.apache.ignite.compute.JobDescriptor;
-import org.apache.ignite.compute.JobExecutionContext;
 import org.apache.ignite.compute.JobTarget;
-import org.apache.ignite.marshaling.ByteArrayMarshaler;
-import org.apache.ignite.marshaling.Marshaler;
+import org.apache.ignite.internal.runner.app.client.Jobs.ArgMarshalingJob;
+import org.apache.ignite.internal.runner.app.client.Jobs.ArgumentAndResultMarshalingJob;
+import org.apache.ignite.internal.runner.app.client.Jobs.ArgumentStringMarshaller;
+import org.apache.ignite.internal.runner.app.client.Jobs.JsonMarshaller;
+import org.apache.ignite.internal.runner.app.client.Jobs.PojoArg;
+import org.apache.ignite.internal.runner.app.client.Jobs.PojoJob;
+import org.apache.ignite.internal.runner.app.client.Jobs.PojoResult;
+import org.apache.ignite.internal.runner.app.client.Jobs.ResultMarshalingJob;
+import org.apache.ignite.internal.runner.app.client.Jobs.ResultStringUnMarshaller;
 import org.apache.ignite.network.ClusterNode;
-import org.jetbrains.annotations.Nullable;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
@@ -175,223 +174,6 @@ public class ItThinClientComputeMarshallingTest extends ItAbstractThinClientTest
                 .addresses(getClientAddresses(List.of(node)).toArray(new String[0]))
                 .build()
                 .compute();
-    }
-
-    static class ArgumentStringMarshaller implements ByteArrayMarshaler<String> {
-        @Override
-        public byte @Nullable [] marshal(@Nullable String object) {
-            return ByteArrayMarshaler.super.marshal(object + ":marshalledOnClient");
-        }
-    }
-
-    static class ArgMarshalingJob implements ComputeJob<String, String> {
-        @Override
-        public CompletableFuture<String> executeAsync(JobExecutionContext context, @Nullable String arg) {
-            return completedFuture(arg + ":processedOnServer");
-        }
-
-        @Override
-        public Marshaler<String, byte[]> inputMarshaler() {
-            return new ByteArrayMarshaler<>() {
-                @Override
-                public String unmarshal(byte @Nullable [] raw) {
-                    return ByteArrayMarshaler.super.unmarshal(raw) + ":unmarshalledOnServer";
-                }
-            };
-        }
-    }
-
-    static class ResultStringUnMarshaller implements ByteArrayMarshaler<String> {
-        @Override
-        public @Nullable String unmarshal(byte @Nullable [] raw) {
-            return ByteArrayMarshaler.super.unmarshal(raw) + ":unmarshalledOnClient";
-        }
-    }
-
-    static class ArgumentAndResultMarshalingJob implements ComputeJob<String, String> {
-        @Override
-        public CompletableFuture<String> executeAsync(JobExecutionContext context, @Nullable String arg) {
-            return completedFuture(arg + ":processedOnServer");
-        }
-
-        @Override
-        public Marshaler<String, byte[]> inputMarshaler() {
-            return new ByteArrayMarshaler<>() {
-                @Override
-                public String unmarshal(byte @Nullable [] raw) {
-                    return ByteArrayMarshaler.super.unmarshal(raw) + ":unmarshalledOnServer";
-                }
-            };
-        }
-
-        @Override
-        public Marshaler<String, byte[]> resultMarshaler() {
-            return new ByteArrayMarshaler<>() {
-                @Override
-                public byte @Nullable [] marshal(@Nullable String object) {
-                    return ByteArrayMarshaler.super.marshal(object + ":marshalledOnServer");
-                }
-            };
-        }
-    }
-
-    static class ResultMarshalingJob implements ComputeJob<String, String> {
-        @Override
-        public CompletableFuture<String> executeAsync(JobExecutionContext context, @Nullable String arg) {
-            return completedFuture(arg + ":processedOnServer");
-        }
-
-        @Override
-        public Marshaler<String, byte[]> resultMarshaler() {
-            return new ByteArrayMarshaler<>() {
-                @Override
-                public byte @Nullable [] marshal(@Nullable String object) {
-                    return ByteArrayMarshaler.super.marshal(object + ":marshalledOnServer");
-                }
-            };
-        }
-    }
-
-    static class JsonMarshalling {
-        private static final ObjectMapper mapper = new ObjectMapper();
-
-        private static byte[] marshal(Object object) {
-            try {
-                return mapper.writeValueAsBytes(object);
-            } catch (JsonProcessingException e) {
-                throw new RuntimeException(e);
-            }
-        }
-
-        private static <T> T unmarshal(byte[] raw, Class<T> clazz) {
-            try {
-                return mapper.readValue(raw, clazz);
-            } catch (IOException e) {
-                throw new RuntimeException(e);
-            }
-        }
-    }
-
-    static class JsonMarshaller<T> implements Marshaler<T, byte[]> {
-        private final Class<T> clazz;
-
-        JsonMarshaller(Class<T> clazz) {
-            this.clazz = clazz;
-        }
-
-        @Override
-        public byte @Nullable [] marshal(@Nullable T object) {
-            return JsonMarshalling.marshal(object);
-        }
-
-        @Override
-        public T unmarshal(byte @Nullable [] raw) {
-            return JsonMarshalling.unmarshal(raw, clazz);
-        }
-    }
-
-    static class PojoJob implements ComputeJob<PojoArg, PojoResult> {
-        @Override
-        public CompletableFuture<PojoResult> executeAsync(JobExecutionContext context, @Nullable PojoArg arg) {
-            var numberFromStr = Integer.parseInt(arg.strValue);
-            return completedFuture(new PojoResult().setLongValue(arg.intValue + numberFromStr));
-        }
-
-        @Override
-        public Marshaler<PojoArg, byte[]> inputMarshaler() {
-            return new JsonMarshaller<>(PojoArg.class);
-        }
-
-        @Override
-        public Marshaler<PojoResult, byte[]> resultMarshaler() {
-            return new JsonMarshaller<>(PojoResult.class);
-        }
-    }
-
-    static class PojoArg {
-        String strValue;
-        int intValue;
-
-        public PojoArg() {
-        }
-
-        public String getStrValue() {
-            return strValue;
-        }
-
-        PojoArg setStrValue(String strValue) {
-            this.strValue = strValue;
-            return this;
-        }
-
-        public int getIntValue() {
-            return intValue;
-        }
-
-        public PojoArg setIntValue(int intValue) {
-            this.intValue = intValue;
-            return this;
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (this == obj) {
-                return true;
-            }
-            if (obj == null || getClass() != obj.getClass()) {
-                return false;
-            }
-            PojoArg pojoArg = (PojoArg) obj;
-            return intValue == pojoArg.intValue && strValue.equals(pojoArg.strValue);
-        }
-
-        @Override
-        public int hashCode() {
-            return strValue.hashCode() + intValue;
-        }
-
-        @Override
-        public String toString() {
-            return "PojoArg{" + "strValue='" + strValue + '\'' + ", intValue=" + intValue + '}';
-        }
-    }
-
-    static class PojoResult {
-        long longValue;
-
-        public PojoResult() {
-        }
-
-        public PojoResult setLongValue(long longValue) {
-            this.longValue = longValue;
-            return this;
-        }
-
-        public long getLongValue() {
-            return longValue;
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (this == obj) {
-                return true;
-            }
-            if (obj == null || getClass() != obj.getClass()) {
-                return false;
-            }
-            PojoResult that = (PojoResult) obj;
-            return longValue == that.longValue;
-        }
-
-        @Override
-        public int hashCode() {
-            return Long.hashCode(longValue);
-        }
-
-        @Override
-        public String toString() {
-            return "PojoResult{" + "longValue=" + longValue + '}';
-        }
     }
 
     private ClusterNode node(int idx) {
